SP2 changes

+++ app/models/sp_staff.rb
@@ -1,4 +1,5 @@
 class SpStaff < ActiveRecord::Base
+  DIRECTORSHIPS = ['PD', 'APD', 'OPD', 'Coordinator']
   unloadable
   set_inheritance_column 'fake_column'
   set_table_name 'sp_staff'
@@ -11,7 +12,7 @@ class SpStaff < ActiveRecord::Base
   
   protected 
     def only_one_of_each_director
-      return true unless %w{PD APD OPD Coordinator}.include?(type)
+      return true unless DIRECTORSHIPS.include?(type)
       SpStaff.where(:type => type, :year => year, :project_id => project_id).first.nil?
     end
     
@@ -26,14 +27,19 @@ class SpStaff < ActiveRecord::Base
         return true if [SpNationalCoordinator, SpRegionalCoordinator].include?(sp_user.class)
         return true if type == 'Evaluator' && sp_user.class == SpDirector
         return true if ['Staff', 'Volunteer'].include?(type)  && sp_user.class == [SpDirector, SpEvaluator].include?(sp_user.class)
-        sp_user.destroy
+        sp_user.delete
       end 
-      base = case type
-             when 'PD', 'APD', 'OPD', 'Coordinator' then SpDirector
-             when 'Evaluator' then SpEvaluator
+      base = case true
+             when DIRECTORSHIPS.include?(type) then SpDirector
+             when type == 'Evaluator' then SpEvaluator
              else SpProjectStaff
              end
-      base.create!(:ssm_id => ssm_id,
-                   :created_at => Time.now)
+      base.create!(:ssm_id => ssm_id, :person_id => person.id)
+    end
+    
+    def destroy_sp_user
+      ssm_id = person.try(:fk_ssmUserId)
+      sp_user = SpUser.where(:ssm_id => ssm_id, :person_id => person.id).first if ssm_id
+      sp_user.destroy if sp_user
     end
 end
